requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.8
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
stripe==11.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
